set exportPackage=false on the js/mixin layer
it used to be set on this layer but the flag got lost
in the creation of html/mixin.  Maybe we should be inheriting
this flag? Also, it broke testUCNav which specifies the full path
name to the type. It seems weird that the layer for the test
script would break that?  When we resolve the absType in processing
a command script, it seems like we should explicitly set the current
model to the package of the type. Or maybe the test script should
be using the package statement before referring to a type so it does
not inherit it?