defined createTreeNode method to workaround constructor problem
- ideally we'd define a constructor in model/TypeNode(TreeEnt ent)
but swing/core/TypeNode(TreeEnt ent) extends a swing class and needs
to call it's constructor. We don't have a way to do super() on both
an extends class and a modified class in this situation (I don't think)

so instead, the model layer uses createTreeNode which is abstract and
overridden in both of those sub-layers.

This actually worked fine before but causes errors in the doc and IDE
occasionally because we were referencing the TreeNode(..) constructor
from the model layer but only defining in the two UI layers downstream.