fixed bug with + sign going away for type tree
- the code to validate the nodes was being called on the client
that did not have the children populated. It would then clear the
flag that went and updated the server. so fixed this twice - don't
do that test on the client at all and don't send back flags from
the client to the server that are not changed on that side